requirements.txt
fastapi==0.115.6
uvicorn==0.30.6
pydantic==2.10.3
pyyaml==6.0.2